requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==4.0.0
django-dotenv==1.4.2
sqlparse==0.4.4
